Setting up the project
