+++ recipes/src/reducers/store.js
@@ -1,0 +1,5 @@
+import { createStore } from 'redux';
+import rootReducer from './reducers';
+
+const store = createStore(rootReducer);
+export default store;
