--- recipes/src/store/reducer.js
@@ -1,8 +1,0 @@
-import data from './data'
-
-const Reducer = (state = data.initialState, action) => {
-  const newState = { ...state };
-  return newState;
-};
-
-export default Reducer;
